requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2

# AI and ML
openai>=1.3.0
anthropic>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.15

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.0
Pillow>=10.0.0
unstructured>=0.10.0
docx2txt>=0.8
pandas>=2.0.0

# Utilities
aiofiles>=23.2.1
python-multipart>=0.0.6
httpx>=0.25.0
tenacity>=8.2.3
tiktoken>=0.5.1

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.10.1
isort>=5.12.0
mypy>=1.6.1
flake8>=6.1.0